Rename profiling module to avoid ambiguity
`mod pprof` was ambiguous in older versions of Rust since we were
also referencing a dependency named `pprof`.
